third_party/blink/web_tests/security/block-test-expected.txt

CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:1/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:7/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:9/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:11/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:13/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:15/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:17/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:19/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:20/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:21/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:22/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:23/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:25/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:37/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:42/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:43/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:53/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:77/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:79/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:87/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:95/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:101/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:102/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:103/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:104/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:109/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:110/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:111/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:113/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:115/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:117/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:119/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:123/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:135/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:139/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:143/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:179/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:389/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:465/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:512/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:513/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:514/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:515/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:526/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:530/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:531/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:532/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:540/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:556/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:563/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:587/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:601/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:636/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:993/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:995/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:2049/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:3659/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:4045/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:6000/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:6665/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:6666/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:6667/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:6668/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:6669/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:65535/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:65536/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:4294967295/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:4294967296/test.jpg
CONSOLE WARNING: Subresource requests using legacy protocols (like `ftp:`) are blocked. Please deliver web-accessible resources over modern protocols like HTTPS. See https://www.chromestatus.com/feature/5709390967472128 for details.
CONSOLE MESSAGE: line 31: PASS: error event fired for ftp://255.255.255.255/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for ftp://255.255.255.255:21/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for ftp://255.255.255.255:22/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:0/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:80/test.jpg
CONSOLE MESSAGE: line 31: PASS: error event fired for http://255.255.255.255:/test.jpg
This test attempts to change the src of an IMG tag to all black listed ports to confirm that each fires an error event. It also tries the FTP ports for exemptions.
